requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
httpx==0.28.1
python-dotenv==1.2.1
openai==2.7.2
pydantic==2.10.6
requests==2.32.5
RapidFuzz==3.14.3
python-dateutil==2.9.0.post0